feat(launchpad): impl FocusManager and move table into separate component - simplifies alot of the old componenet draw/handle key event logic that   used custom `active` field on every component - uses FocusManager to instead deal with which component should be drawn   and should handle keystroke - split the status component into smaller components

+++ node-launchpad/src/config.rs
@@ -118,7 +118,7 @@ pub async fn configure_winsw() -> Result<()> {
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct AppData {
     pub discord_username: String,
-    pub nodes_to_start: usize,
+    pub nodes_to_start: u64,
     pub storage_mountpoint: Option<PathBuf>,
     pub storage_drive: Option<String>,
     pub connection_mode: Option<ConnectionMode>,
